added all parts before training
This code is not optimized. I will improve it in the next days. Area of improvements: 1- Class ReplayMemory can be best implemented by deque, 2- Class DQN's method conv2d_size_out needs to be verified as the formula for the conv output is not as implemented + code can be better.